# backend/msa-account/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.1
  title: Account Management Service
  description: Service that performs account CRUD
  contact:
    name: Byron Delgado
    email: bydelgad@pichincha.com
servers:
  - url: http://localhost:8080
    description: Localhost

paths:
  /cuentas:
    get:
      tags:
        - Account Management
      summary: Obtain the list of accounts
      description: Returns the list of accounts.
      parameters:
        - in: query
          name: accountNumber
          required: false
          description: Account number
          schema:
            type: string
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponseDto'
        "500":
          $ref: "#/components/responses/500"
        "504":
          $ref: "#/components/responses/504"
    post:
      tags:
        - Account Management
      summary: Create a new account
      description: Returns the new account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestDto'
      operationId: Save
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
        "504":
          $ref: "#/components/responses/504"
  /cuentas/{id}:
    get:
      tags:
        - Account Management
      summary: Obtain a account by accountId
      description: Returns a account.
      parameters:
        - name: id
          in: path
          description: account Id
          required: true
          schema:
            type: integer
            example: 1
      operationId: getById
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
        "404":
          $ref: "#/components/responses/404"
        "500":
          $ref: "#/components/responses/500"
        "504":
          $ref: "#/components/responses/504"
    put:
      tags:
        - Account Management
      summary: Update a account
      description: Returns the updated account.
      parameters:
        - name: id
          in: path
          description: account Id
          required: true
          schema:
            type: integer
            example: 1
      operationId: updateById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequestPutDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponseDto'
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
        "504":
          $ref: "#/components/responses/504"
    delete:
      tags:
        - Account Management
      summary: Delete a account by Id
      description: Returns a void
      parameters:
        - name: id
          in: path
          description: account Id
          required: true
          schema:
            type: integer
            example: 1
      operationId: deleteById
      responses:
        "204":
          description: No Content
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
        "504":
          $ref: "#/components/responses/504"
components:
  responses:
    "400":
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsDto"
          example:
            title: Missing mandatory fields.
            detail: The input fields are not correct
            errors:
              - message: Bad Request
                businessMessage: The xyz field is mandatory
            instance: "022"
            type: /v1/name/pathSuffix
    "404":
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsDto"
          example:
            title: Not Found
            detail: Path not found
            instance: "404"
            type: /v1/name/pathSuffix
    "500":
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsDto"
          example:
            title: Internal Server
            detail: An internal error occurred, we are reviewing it.
            instance: "0000"
            type: /v1/name/pathSuffix
            #resource: /xx/v1/yyy
    "504":
      description: Gateway Timeout
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorsDto"
          example:
            title: Gateway Timeout. Request timeout exceeded. If it happens repeatedly, consider reducing the request complexity.
            detail: Gateway connection problems.
            instance: "504"
            type: /v1/name/pathSuffix
  schemas:
    AccountResponseDto:
      type: object
      properties:
        accountId:
          type: integer
          description: Account Id
          nullable: false
          example: 1
        number:
          type: string
          description: Number Account
          nullable: false
          example: 11111111111
        type:
          type: string
          description: Account Type
          nullable: false
          enum:
            - Ahorros
            - Corriente
          example: Ahorros
        balance:
          type: number
          format: double
          description: Balance
          minimum: 0
          nullable: false
          example: 0
        status:
          type: boolean
          description: Status
          nullable: false
          example: true
    AccountRequestDto:
      type: object
      properties:
        type:
          type: string
          description: AccountType
          nullable: false
          enum:
            - Ahorros
            - Corriente
          example: "Ahorros"
        balance:
          type: number
          format: double
          description: Balance
          nullable: false
          example: 0
        clientIdentification:
          type: string
          description: Client Identification
          nullable: false
          example: "1751592013"
      required:
        - type
        - balance
        - clientIdentification
    AccountRequestPutDto:
      type: object
      properties:
        type:
          type: string
          description: AccountType
          nullable: false
          enum:
            - Ahorros
            - Corriente
          example: "Ahorros"
        balance:
          type: number
          format: double
          description: Balance
          nullable: false
          example: 0
        status:
          type: boolean
          description: Status
          nullable: false
          example: true
      required:
        - type
        - balance
        - status
    ErrorDto:
      type: object
      properties:
        message:
          type: string
          description: Detail message
          nullable: false
        businessMessage:
          type: string
          description: Business error message
          nullable: false
    ErrorsDto:
      type: object
      properties:
        title:
          type: string
          description: Technical error message
          nullable: false
          example: Missing mandatory fields.
        detail:
          type: string
          description: Message to display to the client
          nullable: false
          example: The input fields are not correct
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDto'
          nullable: true
        instance:
          type: string
          description: Error catalog code
          nullable: false
          example: "022"
        type:
          type: string